pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "litter-robot-notifier"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pylitterbot>=4.0.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "google-cloud-storage>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-asyncio>=0.21.0"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
